Add support for FileScope with cycles in the imports
The implementation is not the most efficient as it can iterate a lot
before reaching a fixed point, but ultimately the filescope should not
contains too many objects and the iteration should be fast to converge.

But we will need performance testing to see if there is an speed impact.